Create ng2-cache.module.ts
Create separate module for `ng2-cache`.
Without this, the user consuming this package has to mention this as their own providers in *`@NgModule`* (or as `providers` in component metadata).
When we create separate module for CacheService as `Ng2CacheModule` and import, we don't have to do this. This elucidates it more clearly as an external node module/npm package.